move pipeline_E tests to smoke
- reason: these communicate with blazor server.
- todo: also they are adding 30s to runtime for some reason  #82
- success: 65 pass in 40s
